Working on the idea behind SKU segmentation.

+++ backend/inventorysimulator/src/main/java/com/retailops/inventorysimulator/simulator/dto/AbcItemDto.java
@@ -2,10 +2,12 @@ package com.retailops.inventorysimulator.simulator.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @RequiredArgsConstructor
 public class AbcItemDto {
     private String productName;
-    private double salesValue;
+    private BigDecimal salesValue;
 }
